Moved adding/removing roles functionality through reactions  to a separate class called ReactionHandler
