+++ package.json
@@ -12,7 +12,7 @@
     "@discordjs/opus": "^0.9.0",
     "@discordjs/voice": "^0.16.0",
     "axios": "^0.27.2",
-    "discord.js": "^14.11.0",
+    "discord.js": "^14.15.3",
     "ffmpeg-static": "^5.1.0",
     "libsodium-wrappers": "^0.7.11",
     "nodemon": "^2.0.4",
