+++ server.js
@@ -9,6 +9,7 @@ const {
   VoiceChannelChangeHandler,
 } = require('./src/handlers/VoiceChannelChangeHandler');
 const { NewMemberAddHandler } = require('./src/handlers/NewMemberAddHandler');
+const { ReactionHandler } = require('./src/handlers/ReactionHandler');
 
 
 
@@ -31,6 +32,7 @@ const startBot = async () => {
   const messageHandler = new MessageHandler(prefix);
   const vcChangeHandler = new VoiceChannelChangeHandler();
   const newMemberHandler = new NewMemberAddHandler();
+  const reactionHandler = new ReactionHandler();
   client.once('ready', () => {
     messageHandler.initialize();
     client.user.setPresence({
@@ -55,40 +57,11 @@ const startBot = async () => {
   });
 
   client.on('messageReactionAdd', async (reaction, user) => {
-
-    if (REACTIONS[reaction.message.id]) {
-      const emoji = reaction._emoji.name;
-      const member = reaction.message.guild.members.cache.get(user.id);
-      try {
-        const role = reaction.message.guild.roles.cache.get(
-          REACTIONS[reaction.message.id][emoji]
-        );
-        if (REACTIONS[reaction.message.id].reverse) {
-          return member.roles.remove(role);
-        }
-        member.roles.add(role);
-      } catch (e) {
-        console.log('Role does not exist', e);
-      }
-    }
+    reactionHandler.handleReaction(reaction, user);
   });
 
   client.on('messageReactionRemove', async (reaction, user) => {
-    const emoji = reaction._emoji.name;
-    const member = reaction.message.guild.members.cache.get(user.id);
-    if (REACTIONS[reaction.message.id]) {
-      try {
-        const role = reaction.message.guild.roles.cache.get(
-          REACTIONS[reaction.message.id][emoji]
-        );
-        if (REACTIONS[reaction.message.id].reverse) {
-          return member.roles.add(role);
-        }
-        member.roles.remove(role);
-      } catch (e) {
-        console.log('Role does not exist', e);
-      }
-    }
+    reactionHandler.handleReaction(reaction, user, true);
   });
 
   client.login(API_TOKEN);
